# docs/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Business Systems Integration Hub API",
    "version": "1.0.0",
    "description": "A comprehensive middleware platform for synchronizing data between various business systems",
    "contact": {
      "name": "API Support",
      "email": "support@integrationhub.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "CredentialMetadata": {
        "type": "object",
        "properties": {
          "systemType": {
            "type": "string",
            "description": "Type of the system (NetSuite, Salesforce, etc.)"
          },
          "systemId": {
            "type": "string",
            "description": "Unique identifier for the system instance"
          },
          "credentialType": {
            "type": "string",
            "description": "Type of credentials (oauth1, oauth2, basic, api_key)"
          },
          "lastRotated": {
            "type": "string",
            "format": "date-time",
            "description": "When credentials were last rotated"
          },
          "rotationRequired": {
            "type": "boolean",
            "description": "Whether credentials need rotation"
          },
          "accessCount": {
            "type": "number",
            "description": "Number of times credentials have been accessed"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/credentials": {
      "get": {
        "summary": "List all stored credentials (metadata only)",
        "description": "Returns metadata for all stored credentials without exposing sensitive data",
        "tags": [
          "Credentials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of credential metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CredentialMetadata"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      },
      "post": {
        "summary": "Store new system credentials",
        "description": "Securely store credentials for a system using the configured secret manager",
        "tags": [
          "Credentials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "systemType",
                  "systemId",
                  "credentials"
                ],
                "properties": {
                  "systemType": {
                    "type": "string",
                    "description": "Type of system (NetSuite, Salesforce, etc.)",
                    "example": "NetSuite"
                  },
                  "systemId": {
                    "type": "string",
                    "description": "Unique identifier for the system instance",
                    "example": "production"
                  },
                  "credentials": {
                    "type": "object",
                    "description": "Credential object (structure depends on system type)",
                    "example": {
                      "accountId": "12345",
                      "consumerKey": "abcd1234",
                      "consumerSecret": "secret123",
                      "tokenId": "token456",
                      "tokenSecret": "tokensecret789"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Credentials stored successfully"
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      }
    },
    "/api/credentials/{systemType}/{systemId}": {
      "get": {
        "summary": "Get credential metadata for a specific system",
        "description": "Returns metadata for credentials without exposing sensitive data",
        "tags": [
          "Credentials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "systemType",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Type of system"
          },
          {
            "in": "path",
            "name": "systemId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "System identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Credential metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CredentialMetadata"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Credentials not found"
          }
        }
      },
      "delete": {
        "summary": "Delete system credentials",
        "description": "Remove credentials from secure storage",
        "tags": [
          "Credentials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "systemType",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Type of system"
          },
          {
            "in": "path",
            "name": "systemId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "System identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Credentials deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Credentials not found"
          }
        }
      }
    },
    "/api/credentials/rotate": {
      "post": {
        "summary": "Rotate system credentials",
        "description": "Replace existing credentials with new ones, keeping backup",
        "tags": [
          "Credentials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "systemType",
                  "systemId",
                  "newCredentials"
                ],
                "properties": {
                  "systemType": {
                    "type": "string"
                  },
                  "systemId": {
                    "type": "string"
                  },
                  "newCredentials": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credentials rotated successfully"
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Insufficient permissions"
          }
        }
      }
    },
...
    "/api/credentials/security/validate": {
      "get": {
        "summary": "Validate credential security across all integrations",
        "description": "Performs security audit of all integration credentials",
        "tags": [
          "Credentials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Security validation results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalIntegrations": {
                          "type": "number"
                        },
                        "secureIntegrations": {
                          "type": "number"
                        },
                        "insecureIntegrations": {
                          "type": "number"
                        },
                        "issues": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "integrationId": {
                                "type": "string"
                              },
                              "issue": {
                                "type": "string"
                              },
                              "severity": {
                                "type": "string",
                                "enum": [
                                  "high",
                                  "medium",
                                  "low"
                                ]
                              },
                              "recommendation": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/credentials/migrate": {
      "post": {
        "summary": "Migrate credentials to secure storage",
        "description": "Migrate existing credentials from environment variables and inline configs to secret manager",
        "tags": [
          "Credentials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Migration completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "migratedIntegrations": {
                          "type": "number"
                        },
                        "migratedCredentials": {
                          "type": "number"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/credentials/integrations/{integrationId}/credentials": {
      "put": {
        "summary": "Update credentials for an integration",
        "description": "Update credentials for a specific system within an integration",
        "tags": [
          "Credentials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "integrationId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Integration identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "systemType",
                  "newCredentials"
                ],
                "properties": {
                  "systemType": {
                    "type": "string",
                    "enum": [
                      "source",
                      "target"
                    ]
                  },
                  "newCredentials": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Integration credentials updated"
          },
          "400": {
            "description": "Invalid request data"
          },
          "404": {
            "description": "Integration not found"
          }
        }
      }
    },
...
  },
...
}